Continue work on refactoring common module to be modloader-independent, move commands to separate class
