+++ fabric/src/main/java/xyz/eclipseisoffline/eclipsescustomname/fabric/CustomNameFabric.java
@@ -1,20 +1,21 @@
 package xyz.eclipseisoffline.eclipsescustomname.fabric;
 
-import me.lucko.fabric.api.permissions.v0.Permissions;
+import com.mojang.brigadier.CommandDispatcher;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.commands.CommandSourceStack;
-import net.minecraft.server.permissions.Permission;
-import net.minecraft.server.permissions.PermissionLevel;
 import xyz.eclipseisoffline.eclipsescustomname.CustomName;
+import xyz.eclipseisoffline.eclipsescustomname.CustomNamePermissions;
+
+import java.nio.file.Path;
+import java.util.function.Consumer;
 
 public class CustomNameFabric extends CustomName implements ModInitializer {
 
     @Override
     public void onInitialize() {
         initialize();
-        CommandRegistrationCallback.EVENT.register((dispatcher, _, _) -> registerCommands(dispatcher));
     }
 
     @Override
@@ -23,9 +24,17 @@ public class CustomNameFabric extends CustomName implements ModInitializer {
     }
 
     @Override
-    protected boolean hasPermission(CommandSourceStack source, String permission, PermissionLevel fallback) {
-        //return Permissions.check(source, permission, fallback);
-        return source.permissions().hasPermission(new Permission.HasCommandLevel(fallback));
-        return super.hasPermission(source, permission, fallback);
+    protected Path getConfigDir() {
+        return FabricLoader.getInstance().getConfigDir();
+    }
+
+    @Override
+    protected CustomNamePermissions createPermissions() {
+        return new CustomNameFabricPermissions();
+    }
+
+    @Override
+    protected void registerCommands(Consumer<CommandDispatcher<CommandSourceStack>> registerer) {
+        CommandRegistrationCallback.EVENT.register((dispatcher, _, _) -> registerer.accept(dispatcher));
     }
 }
